Refactor navigation structure by removing Navbar and Sidebar components
- Deleted the `NavbarComponent` and `SidebarComponent` files to streamline the project structure.
- Updated import paths in `ClientLayout.tsx` to reflect the removal of these components.
